requirements.txt
# DeepDive Analyst 项目依赖
# AI Agent 框架
crewai>=0.28.0
langgraph>=0.2.0
langchain>=0.3.0

# LLM 提供商支持
langchain-openai>=0.2.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
dashscope>=1.14.0

# 搜索和网络工具
tavily-python>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# CLI 界面
typer>=0.12.0
rich>=13.7.0

# 环境配置
python-dotenv>=1.0.0

# 日志和调试
loguru>=0.7.0

# 测试框架
pytest>=8.0.0
pytest-asyncio>=0.23.0

# 开发工具
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
